ML/models/requirements.txt
flask
flask-cors
tensorflow-cpu
librosa
numpy
boto3
python-dotenv
Pillow
gunicorn